Refine odps_ddl_runner.py and hive_udtf_sql_runner.py
